Made a lot of changes in this one.
Added in the fmtlib formatting library for use in the console driver program for output. Planning to add it for logging in most places.

Added in Configuration class in anticipation of different version variations in SWN, as well as to allow for custom values by the user. Currently only implemented in the FactionControl class for default faction count.

Moved the DiceRoll class and methods into their own independent library to reduce dependency,

Removed the DTO library and changed it to Models, with each Model now a simple includable header file to reduce complexity.

Upgraded Catch library header to v2.0.1

Upgraded C++ version to 14